Restructure files again to make it clearer what is for unit tests and what is for the api, add eslint warnings for no exports
