+++ src/typescript/api/package.json
@@ -8,7 +8,7 @@
   "devDependencies": {
     "@aptos-labs/aptos-cli": "^0.1.2",
     "@types/jest": "^29.5.11",
-    "@types/node": "^20.10.4",
+    "@types/node": "^20.12.7",
     "@typescript-eslint/eslint-plugin": "^6.14.0",
     "@typescript-eslint/parser": "^6.14.0",
     "dotenv": "^16.3.1",
