--- src/typescript/api/src/cli/index.ts
@@ -1,3 +1,0 @@
-export * from "./local-node";
-export * from "./publish";
-export * from "./types";
