+++ src/typescript/api/src/emojicoin_dot_fun/const.ts
