+++ src/typescript/api/src/emojicoin_dot_fun/emojicoin-dot-fun.ts
@@ -19,7 +19,7 @@ import {
   type UserTransactionResponse,
   type MoveValue,
 } from "@aptos-labs/ts-sdk";
-import { type TypeTagInput, type Uint8, type Uint64 } from "./types";
+import { type TypeTagInput, type Uint8, type Uint64 } from ".";
 import {
   ViewFunctionPayloadBuilder,
   EntryFunctionPayloadBuilder,
@@ -59,7 +59,7 @@ export class ProvideLiquidity extends EntryFunctionPayloadBuilder {
   public readonly feePayer?: AccountAddress;
 
   private constructor(args: {
-    moduleAddress: AccountAddressInput; // The module address. You can turn this option off in your config.yaml.
+    moduleAddress: AccountAddressInput;
     marketAddress: AccountAddressInput; // &signer
     provider: AccountAddressInput; // address
     quoteAmount: Uint64; // u64
@@ -164,7 +164,7 @@ export class RegisterMarket extends EntryFunctionPayloadBuilder {
   public readonly feePayer?: AccountAddress;
 
   private constructor(args: {
-    moduleAddress: AccountAddressInput; // The module address. You can turn this option off in your config.yaml.
+    moduleAddress: AccountAddressInput;
     registrant: AccountAddressInput; // &signer
     emojis: Array<HexInput>; // vector<vector<u8>>
     integrator: AccountAddressInput; // address
@@ -263,7 +263,7 @@ export class RemoveLiquidity extends EntryFunctionPayloadBuilder {
   public readonly feePayer?: AccountAddress;
 
   private constructor(args: {
-    moduleAddress: AccountAddressInput; // The module address. You can turn this option off in your config.yaml.
+    moduleAddress: AccountAddressInput;
     marketAddress: AccountAddressInput; // &signer
     provider: AccountAddressInput; // address
     lpCoinAmount: Uint64; // u64
@@ -374,7 +374,7 @@ export class Swap extends EntryFunctionPayloadBuilder {
   public readonly feePayer?: AccountAddress;
 
   private constructor(args: {
-    moduleAddress: AccountAddressInput; // The module address. You can turn this option off in your config.yaml.
+    moduleAddress: AccountAddressInput;
     marketAddress: AccountAddressInput; // &signer
     swapper: AccountAddressInput; // address
     inputAmount: Uint64; // u64
@@ -493,7 +493,7 @@ export class IsASupportedEmoji extends ViewFunctionPayloadBuilder<[boolean]> {
   public readonly typeTags: [] = [];
 
   constructor(args: {
-    moduleAddress: string; // The module address. You can turn this option off in your config.yaml.
+    moduleAddress: string;
 
     hexBytes: HexInput; // vector<u8>
   }) {
@@ -532,7 +532,7 @@ export class IsSupportedEmojiSequence extends ViewFunctionPayloadBuilder<[boolea
   public readonly typeTags: [] = [];
 
   constructor(args: {
-    moduleAddress: string; // The module address. You can turn this option off in your config.yaml.
+    moduleAddress: string;
 
     emojis: Array<HexInput>; // vector<vector<u8>>
   }) {
@@ -575,7 +575,7 @@ export class SimulateProvideLiquidity extends ViewFunctionPayloadBuilder<[MoveVa
   public readonly typeTags: [] = [];
 
   constructor(args: {
-    moduleAddress: string; // The module address. You can turn this option off in your config.yaml.
+    moduleAddress: string;
 
     marketAddress: string; // address
     provider: string; // address
@@ -622,7 +622,7 @@ export class SimulateRemoveLiquidity extends ViewFunctionPayloadBuilder<[MoveVal
   public readonly typeTags: [TypeTag]; // [Emojicoin]
 
   constructor(args: {
-    moduleAddress: string; // The module address. You can turn this option off in your config.yaml.
+    moduleAddress: string;
 
     marketAddress: string; // address
     provider: string; // address
@@ -679,7 +679,7 @@ export class SimulateSwap extends ViewFunctionPayloadBuilder<[MoveValue]> {
   public readonly typeTags: [] = [];
 
   constructor(args: {
-    moduleAddress: string; // The module address. You can turn this option off in your config.yaml.
+    moduleAddress: string;
 
     marketAddress: string; // address
     swapper: string; // address
