+++ src/typescript/api/src/emojicoin_dot_fun/index.ts
@@ -1,1 +1,4 @@
 export * as EmojicoinDotFun from "./emojicoin-dot-fun";
+export * from "./const";
+export * from "./types";
+export * from "./utils";
