+++ src/typescript/api/src/emojicoin_dot_fun/utils.ts
@@ -9,7 +9,7 @@ import {
   DeriveScheme,
 } from "@aptos-labs/ts-sdk";
 import { sha3_256 } from "@noble/hashes/sha3";
-import { EMOJICOIN_DOT_FUN_MODULE_NAME } from "../utils";
+import { EMOJICOIN_DOT_FUN_MODULE_NAME } from "./const";
 
 /**
  * Sleep the current thread for the given amount of time
