--- src/typescript/api/src/helpers/index.ts
@@ -1,3 +1,0 @@
-export * from "./aptos-client";
-export * from "./fund-accounts";
-export * from "./misc";
