+++ src/typescript/api/src/index.ts
@@ -1,4 +1,1 @@
-export * from "./cli";
 export * from "./emojicoin_dot_fun";
-export * from "./helpers";
-export * from "./utils";
