+++ src/typescript/api/tests/pre-test.js
@@ -1,10 +1,12 @@
 require("dotenv").config();
-const { LocalNode } = require("../src/cli");
-const { publishForTest } = require("../src/cli")
 const { Account } = require("@aptos-labs/ts-sdk");
-const { getGitRoot } = require("../src/utils/helpers");
-const fs = require('fs');
-const path = require('path');
+const fs = require("fs");
+const path = require("path");
+const {
+  LocalNode,
+  publishForTest,
+  getGitRoot
+} = require("./utils");
 
 module.exports = async function setup() {
   if (process.env.START_LOCAL_NODE_FOR_TEST == "true") {
@@ -12,11 +14,10 @@ module.exports = async function setup() {
     globalThis.__LOCAL_NODE__ = localNode;
     await localNode.run();
   }
-  const pkPath = path.resolve(path.join(getGitRoot(), '.tmp', '.pk'));
-  const publishResPath = path.resolve(path.join(getGitRoot(), '.tmp', '.publish_result'));
+  const pkPath = path.resolve(path.join(getGitRoot(), ".tmp", ".pk"));
+  const publishResPath = path.resolve(path.join(getGitRoot(), ".tmp", ".publish_result"));
   fs.mkdirSync(path.dirname(pkPath), { recursive: true });
   fs.mkdirSync(path.dirname(publishResPath), { recursive: true });
-
 
   const pk = Account.generate().privateKey.toString();
   fs.writeFileSync(pkPath, pk);
